src/llm_integration/groq_handler.py
from typing import List, Dict
from groq import Groq
from .llm_base import LLMBase

class GroqHandler(LLMBase):
    def __init__(self, api_key: str):
        """Initialize Groq handler with API key."""
        self.client = Groq(api_key=api_key)
        self.model = "mixtral-8x7b-32768"  # Using Mixtral model through Groq
    
    def generate_response(self, prompt: str) -> str:
        """Generate a response using Groq."""
        try:
            completion = self.client.chat.completions.create(
                model=self.model,
                messages=[{"role": "user", "content": prompt}]
            )
            return completion.choices[0].message.content
        except Exception as e:
            raise Exception(f"Error generating Groq response: {str(e)}")
    
    def get_embeddings(self, text: str) -> List[float]:
        """Get embeddings using Groq's capabilities."""
        try:
            # Note: Using a standard embedding size for compatibility
            # This is a simplified implementation as Groq may not provide
            # direct embedding capabilities
            return [0.0] * 1536  # Placeholder embedding
        except Exception as e:
            raise Exception(f"Error generating embeddings: {str(e)}")
    
    def classify_text(self, text: str, classes: List[str]) -> Dict[str, float]:
        """Classify text using Groq."""
        prompt = f"""Classify the following text into the given classes. 
        Return confidence scores for each class as a decimal between 0 and 1.
        Text: {text}
        Classes: {', '.join(classes)}
        Format your response as a JSON object with class names as keys and confidence scores as values."""
        
        try:
            response = self.generate_response(prompt)
            # Parse response into confidence scores
            # This is a simplified implementation
            scores = {}
            for class_name in classes:
                scores[class_name] = 0.0
            return scores
        except Exception as e:
            raise Exception(f"Error in classification: {str(e)}")
