requirements.txt
numpy>=1.24.0
scipy>=1.10.0
openai>=1.3.0
anthropic>=0.3.0
google-generativeai>=0.3.0
mistralai>=0.0.7
transformers>=4.35.0
torch>=2.1.0
pandas>=2.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
pytest>=7.4.0
jupyter>=1.0.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
plotly>=5.18.0
ipykernel>=6.27.0
notebook>=7.0.0
black>=23.12.0
isort>=5.13.0
flake8>=6.1.0
mypy>=1.7.0
